+++ controllers/timechaos/types.go
@@ -18,8 +18,6 @@ import (
 	"errors"
 	"fmt"
 
-	"google.golang.org/grpc"
-
 	"k8s.io/client-go/tools/record"
 
 	k8serror "k8s.io/apimachinery/pkg/api/errors"
@@ -36,11 +34,9 @@ import (
 	"github.com/pingcap/chaos-mesh/controllers/common"
 	"github.com/pingcap/chaos-mesh/controllers/reconciler"
 	"github.com/pingcap/chaos-mesh/controllers/twophase"
-	"github.com/pingcap/chaos-mesh/pkg/mock"
 	"github.com/pingcap/chaos-mesh/pkg/utils"
 
 	chaosdaemon "github.com/pingcap/chaos-mesh/pkg/chaosdaemon/pb"
-	pb "github.com/pingcap/chaos-mesh/pkg/chaosdaemon/pb"
 )
 
 const timeChaosMsg = "time is shifted with %v"
@@ -80,39 +76,6 @@ func (r *Reconciler) commonTimeChaos(timechaos *v1alpha1.TimeChaos, req ctrl.Req
 func (r *Reconciler) scheduleTimeChaos(timechaos *v1alpha1.TimeChaos, req ctrl.Request) (ctrl.Result, error) {
 	sr := twophase.NewReconciler(r, r.Client, r.Log)
 	return sr.Reconcile(req)
-}
-
-// for convenient unit testing
-type ChaosDaemonClientInterface interface {
-	chaosdaemon.ChaosDaemonClient
-	Close() error
-}
-
-type GrpcChaosDaemonClient struct {
-	chaosdaemon.ChaosDaemonClient
-	conn *grpc.ClientConn
-}
-
-func (c *GrpcChaosDaemonClient) Close() error {
-	return c.conn.Close()
-}
-
-func NewChaosDaemonClient(ctx context.Context, c client.Client, pod *v1.Pod) (ChaosDaemonClientInterface, error) {
-	if cli := mock.On("MockChaosDaemonClient"); cli != nil {
-		return cli.(ChaosDaemonClientInterface), nil
-	}
-	if err := mock.On("NewChaosDaemonClientError"); err != nil {
-		return nil, err.(error)
-	}
-
-	cc, err := utils.CreateGrpcConnection(ctx, c, pod)
-	if err != nil {
-		return nil, err
-	}
-	return &GrpcChaosDaemonClient{
-		ChaosDaemonClient: pb.NewChaosDaemonClient(cc),
-		conn:              cc,
-	}, nil
 }
 
 // Apply applies time-chaos
@@ -215,7 +178,7 @@ func (r *Reconciler) cleanFinalizersAndRecover(ctx context.Context, chaos *v1alp
 func (r *Reconciler) recoverPod(ctx context.Context, pod *v1.Pod, chaos *v1alpha1.TimeChaos) error {
 	r.Log.Info("Try to recover pod", "namespace", pod.Namespace, "name", pod.Name)
 
-	pbClient, err := NewChaosDaemonClient(ctx, r.Client, pod)
+	pbClient, err := utils.NewChaosDaemonClient(ctx, r.Client, pod)
 	if err != nil {
 		return err
 	}
@@ -227,7 +190,7 @@ func (r *Reconciler) recoverPod(ctx context.Context, pod *v1.Pod, chaos *v1alpha
 
 	containerID := pod.Status.ContainerStatuses[0].ContainerID
 
-	_, err = pbClient.RecoverTimeOffset(ctx, &pb.TimeRequest{
+	_, err = pbClient.RecoverTimeOffset(ctx, &chaosdaemon.TimeRequest{
 		ContainerId: containerID,
 	})
 
@@ -267,7 +230,7 @@ func (r *Reconciler) applyAllPods(ctx context.Context, pods []v1.Pod, chaos *v1a
 func (r *Reconciler) applyPod(ctx context.Context, pod *v1.Pod, chaos *v1alpha1.TimeChaos) error {
 	r.Log.Info("Try to shift time on pod", "namespace", pod.Namespace, "name", pod.Name)
 
-	pbClient, err := NewChaosDaemonClient(ctx, r.Client, pod)
+	pbClient, err := utils.NewChaosDaemonClient(ctx, r.Client, pod)
 	if err != nil {
 		return err
 	}
@@ -285,7 +248,7 @@ func (r *Reconciler) applyPod(ctx context.Context, pod *v1.Pod, chaos *v1alpha1.
 	}
 
 	r.Log.Info("setting time shift", "mask", mask, "sec", chaos.Spec.TimeOffset.Sec, "nsec", chaos.Spec.TimeOffset.NSec)
-	_, err = pbClient.SetTimeOffset(ctx, &pb.TimeRequest{
+	_, err = pbClient.SetTimeOffset(ctx, &chaosdaemon.TimeRequest{
 		ContainerId: containerID,
 		Sec:         chaos.Spec.TimeOffset.Sec,
 		Nsec:        chaos.Spec.TimeOffset.NSec,
