+++ pkg/utils/chaosdaemon.go
@@ -1,0 +1,45 @@
+package utils
+
+import (
+	"context"
+
+	"google.golang.org/grpc"
+	v1 "k8s.io/api/core/v1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	chaosdaemon "github.com/pingcap/chaos-mesh/pkg/chaosdaemon/pb"
+	"github.com/pingcap/chaos-mesh/pkg/mock"
+)
+
+// for convenient unit testing
+type ChaosDaemonClientInterface interface {
+	chaosdaemon.ChaosDaemonClient
+	Close() error
+}
+
+type GrpcChaosDaemonClient struct {
+	chaosdaemon.ChaosDaemonClient
+	conn *grpc.ClientConn
+}
+
+func (c *GrpcChaosDaemonClient) Close() error {
+	return c.conn.Close()
+}
+
+func NewChaosDaemonClient(ctx context.Context, c client.Client, pod *v1.Pod) (ChaosDaemonClientInterface, error) {
+	if cli := mock.On("MockChaosDaemonClient"); cli != nil {
+		return cli.(ChaosDaemonClientInterface), nil
+	}
+	if err := mock.On("NewChaosDaemonClientError"); err != nil {
+		return nil, err.(error)
+	}
+
+	cc, err := CreateGrpcConnection(ctx, c, pod)
+	if err != nil {
+		return nil, err
+	}
+	return &GrpcChaosDaemonClient{
+		ChaosDaemonClient: chaosdaemon.NewChaosDaemonClient(cc),
+		conn:              cc,
+	}, nil
+}
